requirements.txt
asgiref==3.8.1
Django==4.2.12
djangorestframework==3.15.1
gunicorn==22.0.0
packaging==24.0
psycopg2==2.9.9
python-decouple==3.8
sqlparse==0.5.0
tzdata==2024.1
